fabprimitives/parser/requirements.txt
# Core dependencies
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
aiofiles>=23.2.0
aiohttp>=3.9.0

# Docling for document processing
docling>=2.0.0

# OCR and image processing
tesserocr>=2.6.0
pillow>=10.1.0

# Data processing
pandas>=2.1.0
numpy>=1.24.0

# Background job processing
celery>=5.3.0
redis>=5.0.0

# Database
asyncpg>=0.29.0
sqlalchemy>=2.0.0

# Utilities
python-multipart>=0.0.6
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# Development and testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
httpx>=0.25.0 